Create generate video response and automatic visual and audio editing process to remove inconsistencies
generate video response and automatic visual and audio editing process to remove inconsistencies.

this includes the code for user to upload video, record live videos by accessing their device's front facing camera, and call requests to the APIs needed to generate the video response.  

in the future, we'll add a special custom version of the video generation and audio generative machine learning and data sets we use to make better videos in as close to real time as possible that are the highest quality possible.